refactor | move Patch.java to src/main/java/org/broadinstitute/dsm/model/patch
