+++ src/main/java/org/broadinstitute/dsm/model/patch/Patch.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsm.model;
+package org.broadinstitute.dsm.model.patch;
 
 import com.google.gson.Gson;
 import lombok.Data;
@@ -7,6 +7,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.ddp.db.SimpleResult;
 import org.broadinstitute.dsm.db.structure.DBElement;
 import org.broadinstitute.dsm.exception.DuplicateException;
+import org.broadinstitute.dsm.model.NameValue;
+import org.broadinstitute.dsm.model.Value;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
