+++ src/main/java/org/broadinstitute/dsm/route/PatchRoute.java
@@ -3,9 +3,8 @@ package org.broadinstitute.dsm.route;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import lombok.NonNull;
-import org.broadinstitute.ddp.handlers.util.Result;
 import org.broadinstitute.dsm.exception.DuplicateException;
-import org.broadinstitute.dsm.model.Patch;
+import org.broadinstitute.dsm.model.patch.Patch;
 import org.broadinstitute.dsm.model.patch.BasePatch;
 import org.broadinstitute.dsm.model.patch.PatchFactory;
 import org.broadinstitute.dsm.security.RequestHandler;
